refactor: move cache to individual file
