+++ runtime/src/flow_job/mod.rs
@@ -1,4 +1,5 @@
 mod block_request;
+mod cache;
 pub mod flow;
 mod node_input_values;
 mod run_to_node;
@@ -7,6 +8,7 @@ pub use block_request::{
     parse_node_downstream, parse_query_block_request, parse_run_block_request,
     RunBlockSuccessResponse,
 };
+pub use cache::get_flow_cache_path;
 pub use flow::{execute_flow_job, FlowJobParameters};
 pub use node_input_values::NodeInputValues;
 pub(crate) use upstream::find_upstream_nodes;
